test(plugins): move CLI fork color cases to their own file
The it.each conversion pushed runtime-agent.test.ts past the max-lines
budget once main's adoption tests landed beside it.
